Converted MP calculations to use doubles (#5904)
* Converted editor MP calculations to use doubles

this uses doubles rather than floats to ensure the tiniest of edits can also represent an increase in MP usage, so that there never needs to be a conversion from floats to doubles in this feature

* Add percentage sign to mutation points bar

to make it really clear to everyone that it is a percentage value

* Converted actions and auto-evo steps to use doubles

for MP

* Fixed some display things in the mutation points bar

* Allow MP to go ever so slightly negative

so that rounded values shown in the GUI aren't misleading

* Fixed popup menu move cost displays

* Added rate limiting to editor error sound playing

* Fixed indents in IMutationStrategy.cs

* Slight code change to rigidity slider handling

but it doesn't solve the issue with it being stuck at 1 MP

+++ src/auto-evo/mutation_strategy/ChangeBehaviorScore.cs
@@ -26,7 +26,7 @@ public class ChangeBehaviorScore : IMutationStrategy<MicrobeSpecies>
     // As it cost no MP the mutation code could just repeat this forever
     public bool Repeatable => false;
 
-    public List<Tuple<MicrobeSpecies, float>>? MutationsOf(MicrobeSpecies baseSpecies, float mp, bool lawk,
+    public List<Tuple<MicrobeSpecies, double>>? MutationsOf(MicrobeSpecies baseSpecies, double mp, bool lawk,
         Random random, BiomeConditions biomeToConsider)
     {
         // TODO: Make random something passed in
