+++ src/auto-evo/mutation_strategy/ModifyEnvironmentalTolerance.cs
@@ -12,7 +12,7 @@ public class ModifyEnvironmentalTolerance : IMutationStrategy<MicrobeSpecies>
     /// </summary>
     public bool Repeatable => false;
 
-    public List<Tuple<MicrobeSpecies, float>>? MutationsOf(MicrobeSpecies baseSpecies, float mp, bool lawk,
+    public List<Tuple<MicrobeSpecies, double>>? MutationsOf(MicrobeSpecies baseSpecies, double mp, bool lawk,
         Random random, BiomeConditions biomeToConsider)
     {
         if (mp <= 0)
@@ -53,11 +53,13 @@ public class ModifyEnvironmentalTolerance : IMutationStrategy<MicrobeSpecies>
 
                 if (score.PerfectTemperatureAdjustment < 0)
                 {
-                    change = Math.Max(score.PerfectTemperatureAdjustment, -maxChange);
+                    // Round here so that we don't accidentally round the max change upwards inflating how much change
+                    // is allowed to happen
+                    change = (float)Math.Max(score.PerfectTemperatureAdjustment, -maxChange);
                 }
                 else
                 {
-                    change = Math.Min(score.PerfectTemperatureAdjustment, maxChange);
+                    change = (float)Math.Min(score.PerfectTemperatureAdjustment, maxChange);
                 }
 
                 newTolerances.PreferredTemperature += change;
@@ -68,7 +70,7 @@ public class ModifyEnvironmentalTolerance : IMutationStrategy<MicrobeSpecies>
             else
             {
                 // Trying to perfect this
-                var maxChange = mp / Constants.TOLERANCE_CHANGE_MP_PER_TEMPERATURE_TOLERANCE;
+                var maxChange = (float)(mp / Constants.TOLERANCE_CHANGE_MP_PER_TEMPERATURE_TOLERANCE);
 
                 // To perfect temperature it needs to be always negative
 #if DEBUG
@@ -175,7 +177,7 @@ public class ModifyEnvironmentalTolerance : IMutationStrategy<MicrobeSpecies>
             }
 #endif
 
-            newTolerances.UVResistance += change;
+            newTolerances.UVResistance += (float)change;
 
             mp -= change * Constants.TOLERANCE_CHANGE_MP_PER_OXYGEN;
             changes = true;
@@ -196,7 +198,7 @@ public class ModifyEnvironmentalTolerance : IMutationStrategy<MicrobeSpecies>
             }
 #endif
 
-            newTolerances.UVResistance += change;
+            newTolerances.UVResistance += (float)change;
 
             mp -= change * Constants.TOLERANCE_CHANGE_MP_PER_UV;
             changes = true;
